# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Shot } from './shot';

export interface ProjectUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 20 */
  sceneText?: string;
  resources?: string;
  budget?: string;
  equipment?: string;
  crewExperience?: string;
  challengeMode?: boolean;
  shots?: Shot[];
  productionNotes?: string;
  locked?: boolean;
  projectType?: string;
  overallNotes?: string;
}