# lib/api-zod/src/generated/types/sceneCreate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SceneCreate {
  /** @minLength 1 */
  title: string;
  sceneNumber: number;
  sceneText: string;
}