# lib/api-zod/src/generated/types/sceneUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';
import type { Shot } from './shot';

export interface SceneUpdate {
  title?: string;
  sceneNumber?: number;
  sceneText?: string;
  resources?: string;
  budget?: string;
  equipment?: string;
  crewExperience?: string;
  challengeMode?: boolean;
  messages?: ChatMessage[];
  shots?: Shot[];
  productionNotes?: string;
  readyToLock?: boolean;
  locked?: boolean;
}