# lib/api-zod/src/generated/types/shootDayCreate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShootDayCreate {
  title: string;
  /** @nullable */
  shootDate: string | null;
}