# lib/api-zod/src/generated/types/shootDayShot.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShootDayShot {
  id: string;
  projectId: string;
  sceneId: string;
  shotId: string;
  position: number;
  status: string;
  /** @nullable */
  plannedSetupTime: string | null;
}