# lib/api-zod/src/generated/types/shootDayUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShootDayShot } from './shootDayShot';

export interface ShootDayUpdate {
  title?: string;
  /** @nullable */
  shootDate?: string | null;
  notes?: string;
  status?: string;
  shots?: ShootDayShot[];
}